Adds additional items to the game (currently only 1, but more will be made in the coming weeks).
# Kevin's Classics

Adds additional items to the game (currently only 1, but more will be made in the coming weeks).

Built off of ThinkInvisible's ClassicItems and TILER2. Special thanks to him for making that and for helping me.
Thanks to Nebulon for making the icon(s) used.

These items ARE NOT items from Risk of Rain. Classic Items uses a card system that's extremely easy for importing models, which is why this mod is dependent on it.

## Items

Artemis' Blessing - A Lunar Item; Increases damage the further you are from a target, but there is a median range from the target where if you are closer than range, you will do reduced damage. (This does not affect damage falloff on characters like Commando. I will probably change this in the future.)

## Changelog

1.0.3 - Artemis' Blessing now reduces damage falloff on bullet and shotgun damage (See config for options)